requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
pydantic==2.5.3
openai>=1.30.0
pymupdf>=1.24.0
gunicorn==21.2.0
